delegate the logic to choose b/w registry and plugin to plugin package
