+++ pkg/plugin/plugin.go
@@ -18,99 +18,7 @@
 
 package plugin
 
-import (
-	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
-
-	"github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/go-plugin"
-)
-
-// RevaPlugin represents the runtime plugin
-type RevaPlugin struct {
-	Plugin interface{}
-	Client *plugin.Client
-}
-
-// Kill kills the plugin process
-func (plug *RevaPlugin) Kill() {
-	plug.Client.Kill()
-}
-
-var handshake = plugin.HandshakeConfig{
-	ProtocolVersion:  1,
-	MagicCookieKey:   "BASIC_PLUGIN",
-	MagicCookieValue: "hello",
-}
-
-func compile(pluginType string, path string) (string, error) {
-	dirname := "/var/tmp/reva"
-	binaryPath := filepath.Join(dirname, "bin", pluginType, filepath.Base(path))
-	command := fmt.Sprintf("go build -o %s %s", binaryPath, path)
-	cmd := exec.Command("bash", "-c", command)
-	err := cmd.Run()
-	if err != nil {
-		return "", err
-	}
-	return binaryPath, nil
-}
-
-// checkDir checks and compiles plugin if the configuration points to a directory.
-func checkDirAndCompile(pluginType, driver string) (string, error) {
-	bin := driver
-	file, err := os.Stat(driver)
-	if err != nil {
-		return "", err
-	}
-	// compile if we point to a package
-	if file.IsDir() {
-		bin, err = compile(pluginType, driver)
-		if err != nil {
-			return "", err
-		}
-	}
-	return bin, nil
-}
-
-// Load loads the plugin using the hashicorp go-plugin system
-func Load(driver string, pluginType string) (*RevaPlugin, error) {
-	bin, err := checkDirAndCompile(pluginType, driver)
-	if err != nil {
-		return nil, err
-	}
-
-	logger := hclog.New(&hclog.LoggerOptions{
-		Name:   "plugin",
-		Output: os.Stdout,
-		Level:  hclog.Trace,
-	})
-
-	client := plugin.NewClient(&plugin.ClientConfig{
-		HandshakeConfig: handshake,
-		Plugins:         PluginMap,
-		Cmd:             exec.Command(bin),
-		AllowedProtocols: []plugin.Protocol{
-			plugin.ProtocolNetRPC,
-		},
-		Logger: logger,
-	})
-
-	rpcClient, err := client.Client()
-	if err != nil {
-		return nil, err
-	}
-
-	raw, err := rpcClient.Dispense(pluginType)
-	if err != nil {
-		return nil, err
-	}
-
-	revaPlugin := &RevaPlugin{
-		Plugin: raw,
-		Client: client,
-	}
-
-	return revaPlugin, nil
+// Plugin is the interface used to configure plugins
+type Plugin interface {
+	Configure(m map[string]interface{}) error
 }
